Address review: select-mode columns, hover titles, leaf IP util
